+++ uefi-raw/src/net.rs
@@ -144,6 +144,27 @@ impl IpAddress {
             v6: Ipv6Address(octets),
         }
     }
+
+    /// Transforms this EFI type to the Rust standard library's type
+    /// [`core::net::IpAddr`].
+    ///
+    /// # Arguments
+    /// - `is_ipv6`: Whether the internal data should be interpreted as IPv6 or
+    ///   IPv4 address.
+    ///
+    /// # Safety
+    /// Callers must ensure that the `v4` field is valid if `is_ipv6` is false,
+    /// and that the `v6` field is valid if `is_ipv6` is true
+    #[must_use]
+    pub unsafe fn into_core_ip_addr(self, is_ipv6: bool) -> core::net::IpAddr {
+        if is_ipv6 {
+            // SAFETY: Caller assumes that the underlying data is initialized.
+            core::net::IpAddr::V6(core::net::Ipv6Addr::from(unsafe { self.v6.octets() }))
+        } else {
+            // SAFETY: Caller assumes that the underlying data is initialized.
+            core::net::IpAddr::V4(core::net::Ipv4Addr::from(unsafe { self.v4.octets() }))
+        }
+    }
 }
 
 impl Debug for IpAddress {
@@ -219,6 +240,17 @@ impl MacAddress {
     #[must_use]
     pub const fn octets(self) -> [u8; 32] {
         self.0
+    }
+
+    /// Tries to interpret the MAC address as normal 6-byte MAC address, as used
+    /// in ethernet.
+    pub fn try_into_ethernet_mac_addr(self) -> Result<[u8; 6], [u8; 32]> {
+        let extra = self.octets()[4..].iter().any(|&x| x != 0);
+        if extra {
+            Err(self.0)
+        } else {
+            Ok(self.octets()[..4].try_into().unwrap())
+        }
     }
 }